Remove DuplicationRelationViewRepository
RISDEV-5912
